fix(portal): move the portal rate limiter into the services layer
`services/portal/reportsSelfService` imported `checkRateLimit` from
`routes/portal/helpers`, and that route module reaches
`routes/auth/helpers` → the services barrel → `commandQueue` → `agentWs`
→ the discovery worker. Any test that loads the real portal-flags module
(`orgPortalSettings.test.ts` does, via `vi.importActual`) therefore pulled
the agent WebSocket and discovery worker into its module graph and died
on `networkBaseline.ts` reading `discoveredAssetTypeEnum` off a partial
schema mock — the Test API red on #4815. A service must not import from
routes.

`services/portal/rateLimit.ts` now owns `PORTAL_USE_REDIS`, the bucket
map, the sweep, `capMapByOldest`, and `checkRateLimit`, byte-for-byte;
`routes/portal/helpers` and `routes/portal/schemas` re-export them so the
auth routes, `clearRateLimitKeys`, and existing tests are unchanged. This
also removes the `reportsSelfService` ↔ `routes/portal/schemas` import
cycle.
